+++ src/main.js
@@ -4,11 +4,11 @@ import { shouldSkipProp, shouldSetValue } from './skip.js'
 
 // Merge error properties.
 export default function setErrorProps(error, props, opts) {
-  const { lowPriority } = normalizeOptions(error, props, opts)
+  const { soft } = normalizeOptions(error, props, opts)
 
   // eslint-disable-next-line fp/no-loops
   for (const propName of Reflect.ownKeys(props)) {
-    mergeProp({ error, props, propName, lowPriority })
+    mergeProp({ error, props, propName, soft })
   }
 
   return error
@@ -21,16 +21,16 @@ export default function setErrorProps(error, props, opts) {
 //  - Error properties should not be removed nor overridden as they contain
 //    useful information
 //  - For the same reason, arrays are concatenated
-const mergeProp = function ({ error, props, propName, lowPriority }) {
+const mergeProp = function ({ error, props, propName, soft }) {
   if (shouldSkipProp(props, propName)) {
     return
   }
 
   const errorValue = error[propName]
   const propsValue = props[propName]
-  const mergedValue = mergeValues(errorValue, propsValue, lowPriority)
+  const mergedValue = mergeValues(errorValue, propsValue, soft)
 
-  if (!shouldSetValue(errorValue, mergedValue, lowPriority)) {
+  if (!shouldSetValue(errorValue, mergedValue, soft)) {
     return
   }