+++ src/main.test-d.ts
@@ -18,34 +18,22 @@ expectError(setErrorProps(undefined, error))
 
 setErrorProps(error, {}, {})
 expectAssignable<Options>({})
-setErrorProps(error, {}, { lowPriority: true })
-expectAssignable<Options>({ lowPriority: true })
-expectError(setErrorProps(error, {}, { lowPriority: 'true' }))
-expectNotAssignable<Options>({ lowPriority: 'true' })
+setErrorProps(error, {}, { soft: true })
+expectAssignable<Options>({ soft: true })
+expectError(setErrorProps(error, {}, { soft: 'true' }))
+expectNotAssignable<Options>({ soft: 'true' })
 
 expectType<1>(setErrorProps({} as Error & { a: 0 }, { a: 1 as const }).a)
 expectType<0>(
-  setErrorProps(
-    {} as Error & { a: 0 },
-    { a: 1 as const },
-    { lowPriority: true },
-  ).a,
+  setErrorProps({} as Error & { a: 0 }, { a: 1 as const }, { soft: true }).a,
 )
 expectType<1>(setErrorProps({} as Error & { a: {} }, { a: 1 as const }).a)
 expectType<{}>(
-  setErrorProps(
-    {} as Error & { a: {} },
-    { a: 1 as const },
-    { lowPriority: true },
-  ).a,
+  setErrorProps({} as Error & { a: {} }, { a: 1 as const }, { soft: true }).a,
 )
 expectType<{}>(setErrorProps({} as Error & { a: 0 }, { a: {} as const }).a)
 expectType<0>(
-  setErrorProps(
-    {} as Error & { a: 0 },
-    { a: {} as const },
-    { lowPriority: true },
-  ).a,
+  setErrorProps({} as Error & { a: 0 }, { a: {} as const }, { soft: true }).a,
 )
 expectType<1>(
   setErrorProps({} as Error & { a: { b: 0 } }, { a: { b: 1 } as const }).a.b,
@@ -54,7 +42,7 @@ expectType<0>(
   setErrorProps(
     {} as Error & { a: { b: 0 } },
     { a: { b: 1 } as const },
-    { lowPriority: true },
+    { soft: true },
   ).a.b,
 )
 expectType<0>(
@@ -64,7 +52,7 @@ expectType<0>(
   setErrorProps(
     {} as Error & { a: { b: 0 } },
     { a: { c: 1 } as const },
-    { lowPriority: true },
+    { soft: true },
   ).a.b,
 )
 expectType<1>(
@@ -74,32 +62,20 @@ expectType<1>(
   setErrorProps(
     {} as Error & { a: { c: 0 } },
     { a: { b: 1 } as const },
-    { lowPriority: true },
+    { soft: true },
   ).a.b,
 )
 expectType<1>(setErrorProps({} as Error & { a: [0] }, { a: 1 as const }).a)
 expectType<[0]>(
-  setErrorProps(
-    {} as Error & { a: [0] },
-    { a: 1 as const },
-    { lowPriority: true },
-  ).a,
+  setErrorProps({} as Error & { a: [0] }, { a: 1 as const }, { soft: true }).a,
 )
 expectType<[1]>(setErrorProps({} as Error & { a: 0 }, { a: [1] as [1] }).a)
 expectType<0>(
-  setErrorProps(
-    {} as Error & { a: 0 },
-    { a: [1] as [1] },
-    { lowPriority: true },
-  ).a,
+  setErrorProps({} as Error & { a: 0 }, { a: [1] as [1] }, { soft: true }).a,
 )
 expectType<[0, 1]>(setErrorProps({} as Error & { a: [0] }, { a: [1] as [1] }).a)
 expectType<[1, 0]>(
-  setErrorProps(
-    {} as Error & { a: [0] },
-    { a: [1] as [1] },
-    { lowPriority: true },
-  ).a,
+  setErrorProps({} as Error & { a: [0] }, { a: [1] as [1] }, { soft: true }).a,
 )
 expectType<1>(
   setErrorProps({} as Error & { a: undefined }, { a: 1 as const }).a,
@@ -108,22 +84,17 @@ expectType<1>(
   setErrorProps(
     {} as Error & { a: undefined },
     { a: 1 as const },
-    { lowPriority: true },
+    { soft: true },
   ).a,
 )
 expectType<1>(setErrorProps({} as Error, { a: 1 as const }).a)
-expectType<1>(
-  setErrorProps({} as Error, { a: 1 as const }, { lowPriority: true }).a,
-)
+expectType<1>(setErrorProps({} as Error, { a: 1 as const }, { soft: true }).a)
 expectType<undefined>(setErrorProps({} as Error & { a: 1 }, { a: undefined }).a)
 expectType<1>(
-  setErrorProps({} as Error & { a: 1 }, { a: undefined }, { lowPriority: true })
-    .a,
+  setErrorProps({} as Error & { a: 1 }, { a: undefined }, { soft: true }).a,
 )
 expectType<1>(setErrorProps({} as Error & { a: 1 }, {}).a)
-expectType<1>(
-  setErrorProps({} as Error & { a: 1 }, {}, { lowPriority: true }).a,
-)
+expectType<1>(setErrorProps({} as Error & { a: 1 }, {}, { soft: true }).a)
 const symbol = Symbol('test')
 expectType<1>(
   setErrorProps({} as Error & { [symbol]: 0 }, { [symbol]: 1 as const })[
@@ -134,7 +105,7 @@ expectType<0>(
   setErrorProps(
     {} as Error & { [symbol]: 0 },
     { [symbol]: 1 as const },
-    { lowPriority: true },
+    { soft: true },
   )[symbol],
 )
 expectType<'0'>(
@@ -144,7 +115,7 @@ expectType<'0'>(
   setErrorProps(
     {} as Error & { name: '0' },
     { name: '1' as const },
-    { lowPriority: true },
+    { soft: true },
   ).name,
 )
 expectAssignable<() => '0'>(
@@ -155,6 +126,6 @@ expectAssignable<() => '0'>(
   setErrorProps(
     {} as Error & { toString: () => '0' },
     { toString: () => '1' },
-    { lowPriority: true },
+    { soft: true },
   ).toString,
 )