+++ src/merge.js
@@ -2,10 +2,10 @@ import deepMerge from 'deepmerge'
 import isPlainObj from 'is-plain-obj'
 
 // Merge `error[propName]` and `props[propName]`
-export const mergeValues = function (errorValue, propsValue, lowPriority) {
+export const mergeValues = function (errorValue, propsValue, soft) {
   return shouldMergeValues(errorValue, propsValue)
-    ? mergeDeepValue(errorValue, propsValue, lowPriority)
-    : mergeFlatValue(errorValue, propsValue, lowPriority)
+    ? mergeDeepValue(errorValue, propsValue, soft)
+    : mergeFlatValue(errorValue, propsValue, soft)
 }
 
 const shouldMergeValues = function (errorValue, propsValue) {
@@ -16,8 +16,8 @@ const isArrayOrObject = function (value) {
   return isPlainObj(value) || Array.isArray(value)
 }
 
-const mergeDeepValue = function (errorValue, propsValue, lowPriority) {
-  return lowPriority
+const mergeDeepValue = function (errorValue, propsValue, soft) {
+  return soft
     ? deepMerge(propsValue, errorValue, DEEP_MERGE_OPTS)
     : deepMerge(errorValue, propsValue, DEEP_MERGE_OPTS)
 }
@@ -28,6 +28,6 @@ const isMergeableObject = function (value) {
 
 const DEEP_MERGE_OPTS = { isMergeableObject }
 
-const mergeFlatValue = function (errorValue, propsValue, lowPriority) {
-  return lowPriority && errorValue !== undefined ? errorValue : propsValue
+const mergeFlatValue = function (errorValue, propsValue, soft) {
+  return soft && errorValue !== undefined ? errorValue : propsValue
 }